Relocating factory to primitives.
