+++ src/primitives/automowerAccessoryFactory.ts
@@ -1,15 +1,15 @@
 import { API, PlatformAccessory } from 'homebridge';
 
-import { PlatformAccessoryFactory } from '../primitives/platformAccessoryFactory';
+import { PlatformAccessoryFactory } from './platformAccessoryFactory';
 import { AutomowerAccessory, AutomowerContext } from '../automowerAccessory';
 import { Mower } from '../model';
-import { AccessoryInformationService, AccessoryInformationServiceImpl } from './homebridge/accessoryInformationService';
-import { BatteryService, BatteryServiceImpl } from './homebridge/batteryService';
-import { ScheduleSwitch, ScheduleSwitchImpl } from './homebridge/scheduleSwitch';
-import { PlatformContainer } from '../primitives/platformContainer';
-import { MowerControlServiceImpl } from './automower/mowerControlService';
+import { AccessoryInformationService, AccessoryInformationServiceImpl } from '../services/homebridge/accessoryInformationService';
+import { BatteryService, BatteryServiceImpl } from '../services/homebridge/batteryService';
+import { ScheduleSwitch, ScheduleSwitchImpl } from '../services/homebridge/scheduleSwitch';
+import { PlatformContainer } from './platformContainer';
+import { MowerControlServiceImpl } from '../services/automower/mowerControlService';
 import { PlatformLogger } from '../diagnostics/platformLogger';
-import { DeterministicScheduleEnabledPolicy } from './homebridge/policies/scheduleEnabledPolicy';
+import { DeterministicScheduleEnabledPolicy } from '../services/homebridge/policies/scheduleEnabledPolicy';
 
 /**
  * A mechanism to create {@link AutomowerAccessory} instances.
