+++ src/services/discoveryService.ts
@@ -2,7 +2,7 @@ import { AutomowerAccessory } from '../automowerAccessory';
 import { AutomowerPlatform } from '../automowerPlatform';
 import { PlatformLogger } from '../diagnostics/platformLogger';
 import { GetMowersService } from './automower/getMowersService';
-import { AutomowerAccessoryFactory } from './automowerAccessoryFactory';
+import { AutomowerAccessoryFactory } from '../primitives/automowerAccessoryFactory';
 
 /**
  * A mechanism to discover any automowers associated with an account.
